fix(platform): Enable scrolling via keyboard buttons (#12460)
* enable scrolling via keyboard buttons

* conditionally add event listeners in search

* undo first try

* refactor
