+++ src/components/search/util.ts
@@ -1,0 +1,7 @@
+import {isDeveloperDocs} from 'sentry-docs/isDeveloperDocs';
+
+export function relativizeUrl(url: string) {
+  return isDeveloperDocs
+    ? url
+    : url.replace(/^(https?:\/\/docs\.sentry\.io)(?=\/|$)/, '');
+}
